requirements.txt
Flask==2.2.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
PyJWT~=2.8.0
Werkzeug~=2.3.6
psycopg2-binary==2.9.10
gunicorn==23.0.0